+++ R/survival_cov_adj.R
@@ -152,16 +152,24 @@ h_get_lm_input <- function(df, model) {
   assert_subset(c("treatment", "O_hat", all.vars(model)), names(df))
   assert_factor(df$treatment)
 
-  # Add outcome, remove intercept:
-  model <- stats::update(model, O_hat ~ . - 1)
-  df_by_trt <- split(df, f = df$treatment)
+  # Add outcome:
+  model <- stats::update(model, O_hat ~ .)
+  includes_intercept <- attr(stats::terms(model), "intercept") == 1
+  assert_true(includes_intercept)
+
+  # Create overall design matrix and response vector.
+  mf <- stats::model.frame(model, data = df)
+  x <- stats::model.matrix(model, data = mf)
+  assert_true(identical(colnames(x)[1L], "(Intercept)"))
+  x <- x[, -1L, drop = FALSE] # Remove intercept.
+  y <- stats::model.response(mf)
+
+  # Split both design matrix and response vector by treatment arm.
+  index_by_trt <- split(seq_along(y), f = df$treatment)
   lapply(
-    df_by_trt,
-    function(this_df) {
-      mf <- stats::model.frame(model, data = this_df)
-      x <- stats::model.matrix(model, data = mf)
-      y <- stats::model.response(mf)
-      list(X = x, y = y)
+    index_by_trt,
+    function(this_index) {
+      list(X = x[this_index, , drop = FALSE], y = y[this_index])
     }
   )
 }