+++ tests-raw/test-survival.R
@@ -73,6 +73,21 @@ robincar_args <- list(
 )
 calc_robin_car(robincar_args)
 
+# test: "robin_surv gives the same results as RobinCar for single factor covariate"
+robincar_args <- list(
+  df = na.omit(surv_data),
+  treat_col = "ecog",
+  response_col = "time",
+  event_col = "status",
+  car_strata_cols = NULL,
+  covariate_cols = "sex", # We use a single factor covariate here.
+  car_scheme = "simple",
+  adj_method = "CL",
+  ref_arm = "0",
+  p_trt = mean(surv_data$ecog == "1", na.rm = TRUE)
+)
+calc_robin_car(robincar_args)
+
 # test: "robin_surv_strata_cov gives the same results as RobinCar functions"
 robincar_args <- list(
   df = na.omit(surv_data),
@@ -85,5 +100,29 @@ robincar_args <- list(
   adj_method = "CSL",
   ref_arm = "0",
   p_trt = mean(surv_data$ecog == "1", na.rm = TRUE)
+)
+calc_robin_car(robincar_args)
+
+# test: "robin_surv_strata_cov gives the same results as RobinCar for strong correlation covariate with strata"
+set.seed(2040)
+surv_data2 <- surv_data
+surv_data2$ecog <- as.factor(ifelse(
+  surv_data$strata == 1,
+  rbinom(nrow(surv_data), 1, prob = 0.2),
+  surv_data$ecog
+))
+surv_data2 <- surv_data2[surv_data2$strata %in% c(0, 1), ]
+
+robincar_args <- list(
+  df = na.omit(surv_data2),
+  treat_col = "sex",
+  response_col = "time",
+  event_col = "status",
+  car_strata_cols = "strata",
+  covariate_cols = "ecog",
+  car_scheme = "permuted-block",
+  adj_method = "CSL",
+  ref_arm = "Female",
+  p_trt = mean(surv_data2$sex == "Male", na.rm = TRUE)
 )
 calc_robin_car(robincar_args)